play around with json/tables
