+++ .lua-format
@@ -5,21 +5,21 @@ break_after_functioncall_lp: false
 break_after_functiondef_lp: false
 break_after_operator: false 
 break_after_table_lb: true
-break_before_functioncall_rp: true
-break_before_functiondef_rp: true
-break_before_table_rb: true
+break_before_functioncall_rp: false
+break_before_functiondef_rp: false
+break_before_table_rb: false
 chop_down_kv_table: true
 chop_down_parameter: true
-chop_down_table: false
+chop_down_table: true
 column_limit: 100
-column_table_limit: 100
+column_table_limit: 80
 continuation_indent_width: 2
 double_quote_to_single_quote: false
-extra_sep_at_table_end: false
+extra_sep_at_table_end: true
 indent_width: 2
-keep_simple_control_block_one_line: true
-keep_simple_function_one_line: true
-single_quote_to_double_quote: false
+keep_simple_control_block_one_line: false
+keep_simple_function_one_line: false
+single_quote_to_double_quote: true
 spaces_before_call: 1
 tab_width: 2
 table_sep: ','
